@@ -4,15 +4,15 @@ import com.koushikdutta.async.callback.CompletedCallback;
 import com.koushikdutta.async.callback.DataCallback;
 
 public interface DataEmitter {
-    public void setDataCallback(DataCallback callback);
-    public DataCallback getDataCallback();
-    public boolean isChunked();
-    public void pause();
-    public void resume();
-    public void close();
-    public boolean isPaused();
-    public void setEndCallback(CompletedCallback callback);
-    public CompletedCallback getEndCallback();
-    public AsyncServer getServer();
-    public String charset();
+    void setDataCallback(DataCallback callback);
+    DataCallback getDataCallback();
+    boolean isChunked();
+    void pause();
+    void resume();
+    void close();
+    boolean isPaused();
+    void setEndCallback(CompletedCallback callback);
+    CompletedCallback getEndCallback();
+    AsyncServer getServer();
+    String charset();
 }
