@@ -4,7 +4,7 @@ package com.koushikdutta.async;
  * Created by koush on 5/28/13.
  */
 public interface DataTrackingEmitter extends DataEmitter {
-    public interface DataTracker {
+    interface DataTracker {
         void onData(int totalBytesRead);
     }
     void setDataTracker(DataTracker tracker);
