@@ -1,5 +1,5 @@
 package com.koushikdutta.async.future;
 
 public interface DependentCancellable extends Cancellable {
-    public DependentCancellable setParent(Cancellable parent);
+    boolean setParent(Cancellable parent);
 }
