@@ -13,51 +13,54 @@ import com.koushikdutta.async.util.UntypedHashtable;
  * inspect, manipulate, and handle http requests.
  */
 public interface AsyncHttpClientMiddleware {
-    public interface ResponseHead  {
-        public AsyncSocket socket();
-        public String protocol();
-        public String message();
-        public int code();
-        public ResponseHead protocol(String protocol);
-        public ResponseHead message(String message);
-        public ResponseHead code(int code);
-        public Headers headers();
-        public ResponseHead headers(Headers headers);
-        public DataSink sink();
-        public ResponseHead sink(DataSink sink);
-        public DataEmitter emitter();
-        public ResponseHead emitter(DataEmitter emitter);
+    interface ResponseHead  {
+        AsyncSocket socket();
+        String protocol();
+        String message();
+        int code();
+        ResponseHead protocol(String protocol);
+        ResponseHead message(String message);
+        ResponseHead code(int code);
+        Headers headers();
+        ResponseHead headers(Headers headers);
+        DataSink sink();
+        ResponseHead sink(DataSink sink);
+        DataEmitter emitter();
+        ResponseHead emitter(DataEmitter emitter);
     }
 
-    public static class OnRequestData {
+    class OnRequestData {
         public UntypedHashtable state = new UntypedHashtable();
         public AsyncHttpRequest request;
     }
 
-    public static class GetSocketData extends OnRequestData {
+    class GetSocketData extends OnRequestData {
         public ConnectCallback connectCallback;
         public Cancellable socketCancellable;
         public String protocol;
     }
 
-    public static class OnExchangeHeaderData extends GetSocketData {
+    class OnExchangeHeaderData extends GetSocketData {
         public AsyncSocket socket;
         public ResponseHead response;
         public CompletedCallback sendHeadersCallback;
         public CompletedCallback receiveHeadersCallback;
     }
 
-    public static class OnRequestSentData extends OnExchangeHeaderData {
+    class OnRequestSentData extends OnExchangeHeaderData {
     }
 
-    public static class OnHeadersReceivedDataOnRequestSentData extends OnRequestSentData {
+    class OnHeadersReceivedData extends OnRequestSentData {
     }
 
-    public static class OnBodyDataOnRequestSentData extends OnHeadersReceivedDataOnRequestSentData {
+    class OnBodyDecoderData extends OnHeadersReceivedData {
         public DataEmitter bodyEmitter;
     }
 
-    public static class OnResponseCompleteDataOnRequestSentData extends OnBodyDataOnRequestSentData {
+    class OnResponseReadyData extends OnBodyDecoderData {
+    }
+
+    class OnResponseCompleteData extends OnResponseReadyData {
         public Exception exception;
     }
 
@@ -65,14 +68,14 @@ public interface AsyncHttpClientMiddleware {
      * Called immediately upon request execution
      * @param data
      */
-    public void onRequest(OnRequestData data);
+    void onRequest(OnRequestData data);
 
     /**
      * Called to retrieve the socket that will fulfill this request
      * @param data
      * @return
      */
-    public Cancellable getSocket(GetSocketData data);
+    Cancellable getSocket(GetSocketData data);
 
     /**
      * Called before when the headers are sent and received via the socket.
@@ -80,31 +83,40 @@ public interface AsyncHttpClientMiddleware {
      * @param data
      * @return
      */
-    public boolean exchangeHeaders(OnExchangeHeaderData data);
+    boolean exchangeHeaders(OnExchangeHeaderData data);
 
     /**
      * Called once the headers and any optional request body has
      * been sent
      * @param data
      */
-    public void onRequestSent(OnRequestSentData data);
+    void onRequestSent(OnRequestSentData data);
 
     /**
      * Called once the headers have been received via the socket
      * @param data
      */
-    public void onHeadersReceived(OnHeadersReceivedDataOnRequestSentData data);
+    void onHeadersReceived(OnHeadersReceivedData data);
 
     /**
      * Called before the body is decoded
      * @param data
      */
-    public void onBodyDecoder(OnBodyDataOnRequestSentData data);
+    void onBodyDecoder(OnBodyDecoderData data);
+
+    /**
+     * Called before the response is returned to the client. Return a new AsyncHttpRequest
+     * to end the current request and start a new one. Can be used to implement redirect strategies
+     * or multileg authentication, such as digest.
+     * @param data
+     * @return
+     */
+    AsyncHttpRequest onResponseReady(OnResponseReadyData data);
 
     /**
      * Called once the request is complete and response has been received,
      * or if an error occurred
      * @param data
      */
-    public void onResponseComplete(OnResponseCompleteDataOnRequestSentData data);
+    void onResponseComplete(OnResponseCompleteData data);
 }
