@@ -32,8 +32,8 @@ public class HttpUtil {
                 if (StringBody.CONTENT_TYPE.equals(ct)) {
                     return new StringBody();
                 }
-                if (MultipartFormDataBody.CONTENT_TYPE.equals(ct)) {
-                    return new MultipartFormDataBody(values);
+                if (ct != null && ct.startsWith(MultipartFormDataBody.PRIMARY_TYPE)) {
+                    return new MultipartFormDataBody(contentType);
                 }
             }
         }
@@ -60,12 +60,13 @@ public class HttpUtil {
     }
     
     public static DataEmitter getBodyDecoder(DataEmitter emitter, Protocol protocol, Headers headers, boolean server) {
-        long _contentLength;
+        long _contentLength = -1;
         try {
-            _contentLength = Long.parseLong(headers.get("Content-Length"));
+            String header = headers.get("Content-Length");
+            if (header != null)
+                _contentLength = Long.parseLong(header);
         }
-        catch (Exception ex) {
-            _contentLength = -1;
+        catch (NumberFormatException ex) {
         }
         final long contentLength = _contentLength;
         if (-1 != contentLength) {
@@ -90,14 +91,12 @@ public class HttpUtil {
             chunker.setDataEmitter(emitter);
             emitter = chunker;
         }
-        else {
-            if ((server || protocol == Protocol.HTTP_1_1) && !"close".equalsIgnoreCase(headers.get("Connection"))) {
-                // if this is the server, and the client has not indicated a request body, the client is done
-                EndEmitter ender = EndEmitter.create(emitter.getServer(), null);
-                ender.setDataEmitter(emitter);
-                emitter = ender;
-                return emitter;
-            }
+        else if (server) {
+            // if this is the server, and the client has not indicated a request body, the client is done
+            EndEmitter ender = EndEmitter.create(emitter.getServer(), null);
+            ender.setDataEmitter(emitter);
+            emitter = ender;
+            return emitter;
         }
 
         if ("gzip".equals(headers.get("Content-Encoding"))) {
@@ -132,12 +131,12 @@ public class HttpUtil {
         return "keep-alive".equalsIgnoreCase(connection);
     }
 
-    public static int contentLength(Headers headers) {
+    public static long contentLength(Headers headers) {
         String cl = headers.get("Content-Length");
         if (cl == null)
             return -1;
         try {
-            return Integer.parseInt(cl);
+            return Long.parseLong(cl);
         }
         catch (NumberFormatException e) {
             return -1;
