@@ -1,10 +1,8 @@
 package com.koushikdutta.async.http.server;
 
 import android.annotation.TargetApi;
-import android.content.Context;
-import android.content.res.AssetManager;
 import android.os.Build;
-import android.text.TextUtils;
+import android.util.Log;
 
 import com.koushikdutta.async.AsyncSSLSocket;
 import com.koushikdutta.async.AsyncSSLSocketWrapper;
@@ -16,34 +14,21 @@ import com.koushikdutta.async.DataEmitter;
 import com.koushikdutta.async.Util;
 import com.koushikdutta.async.callback.CompletedCallback;
 import com.koushikdutta.async.callback.ListenCallback;
-import com.koushikdutta.async.http.AsyncHttpGet;
-import com.koushikdutta.async.http.AsyncHttpHead;
-import com.koushikdutta.async.http.AsyncHttpPost;
+import com.koushikdutta.async.callback.ValueCallback;
 import com.koushikdutta.async.http.Headers;
 import com.koushikdutta.async.http.HttpUtil;
 import com.koushikdutta.async.http.Multimap;
-import com.koushikdutta.async.http.Protocol;
 import com.koushikdutta.async.http.WebSocket;
-import com.koushikdutta.async.http.WebSocketImpl;
 import com.koushikdutta.async.http.body.AsyncHttpRequestBody;
-import com.koushikdutta.async.util.StreamUtility;
 
-import java.io.File;
-import java.io.FileInputStream;
-import java.io.FileNotFoundException;
-import java.io.IOException;
-import java.io.InputStream;
+import java.net.URLDecoder;
 import java.util.ArrayList;
-import java.util.Collections;
-import java.util.Comparator;
 import java.util.Hashtable;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
 
 import javax.net.ssl.SSLContext;
 
 @TargetApi(Build.VERSION_CODES.ECLAIR)
-public class AsyncHttpServer {
+public class AsyncHttpServer extends AsyncHttpServerRouter {
     ArrayList<AsyncServerSocket> mListeners = new ArrayList<AsyncServerSocket>();
     public void stop() {
         if (mListeners != null) {
@@ -57,26 +42,85 @@ public class AsyncHttpServer {
         return false;
     }
 
+    protected void onResponseCompleted(AsyncHttpServerRequest request, AsyncHttpServerResponse response) {
+
+    }
+
     protected void onRequest(HttpServerRequestCallback callback, AsyncHttpServerRequest request, AsyncHttpServerResponse response) {
-        if (callback != null)
-            callback.onRequest(request, response);
+        if (callback != null) {
+            try {
+                callback.onRequest(request, response);
+            }
+            catch (Exception e) {
+                Log.e("AsyncHttpServer", "request callback raised uncaught exception. Catching versus crashing process", e);
+                response.code(500);
+                response.end();
+            }
+        }
+    }
+
+    protected boolean isKeepAlive(AsyncHttpServerRequest request, AsyncHttpServerResponse response) {
+        return HttpUtil.isKeepAlive(response.getHttpVersion(), request.getHeaders());
     }
 
     protected AsyncHttpRequestBody onUnknownBody(Headers headers) {
         return new UnknownRequestBody(headers.get("Content-Type"));
     }
 
+    protected boolean isSwitchingProtocols(AsyncHttpServerResponse res) {
+        return res.code() == 101;
+    }
+
     ListenCallback mListenCallback = new ListenCallback() {
         @Override
         public void onAccepted(final AsyncSocket socket) {
-            AsyncHttpServerRequestImpl req = new AsyncHttpServerRequestImpl() {
-                HttpServerRequestCallback match;
+            final AsyncHttpServerRequestImpl req = new AsyncHttpServerRequestImpl() {
+                AsyncHttpServerRequestImpl self = this;
+                HttpServerRequestCallback requestCallback;
                 String fullPath;
                 String path;
                 boolean responseComplete;
                 boolean requestComplete;
                 AsyncHttpServerResponseImpl res;
                 boolean hasContinued;
+                boolean handled;
+
+                final Runnable onFinally = new Runnable() {
+                    @Override
+                    public void run() {
+                        Log.i("HTTP", "Done");
+                    }
+                };
+
+                final ValueCallback<Exception> onException = new ValueCallback<Exception>() {
+                    @Override
+                    public void onResult(Exception value) {
+                        Log.e("HTTP", "exception", value);
+                    }
+                };
+
+                void onRequest() {
+                    AsyncHttpServer.this.onRequest(requestCallback, this, res);
+                }
+
+                @Override
+                protected AsyncHttpRequestBody onBody(Headers headers) {
+                    String statusLine = getStatusLine();
+                    String[] parts = statusLine.split(" ");
+                    fullPath = parts[1];
+                    path = URLDecoder.decode(fullPath.split("\\?")[0]);
+                    method = parts[0];
+                    RouteMatch route = route(method, path);
+                    if (route == null)
+                        return null;
+
+                    matcher = route.matcher;
+                    requestCallback = route.callback;
+
+                    if (route.bodyCallback == null)
+                        return null;
+                    return route.bodyCallback.getBody(headers);
+                }
 
                 @Override
                 protected AsyncHttpRequestBody onUnknownBody(Headers headers) {
@@ -108,24 +152,6 @@ public class AsyncHttpServer {
                     }
 //                    System.out.println(headers.toHeaderString());
                     
-                    String statusLine = getStatusLine();
-                    String[] parts = statusLine.split(" ");
-                    fullPath = parts[1];
-                    path = fullPath.split("\\?")[0];
-                    method = parts[0];
-                    synchronized (mActions) {
-                        ArrayList<Pair> pairs = mActions.get(method);
-                        if (pairs != null) {
-                            for (Pair p: pairs) {
-                                Matcher m = p.regex.matcher(path);
-                                if (m.matches()) {
-                                    mMatcher = m;
-                                    match = p.callback;
-                                    break;
-                                }
-                            }
-                        }
-                    }
                     res = new AsyncHttpServerResponseImpl(socket, this) {
                         @Override
                         protected void report(Exception e) {
@@ -139,35 +165,36 @@ public class AsyncHttpServer {
 
                         @Override
                         protected void onEnd() {
+                            responseComplete = true;
                             super.onEnd();
                             mSocket.setEndCallback(null);
-                            responseComplete = true;
+
+                            onResponseCompleted(getRequest(), res);
+
                             // reuse the socket for a subsequent request.
                             handleOnCompleted();
                         }
                     };
-                    
-                    boolean handled = onRequest(this, res);
 
-                    if (match == null && !handled) {
+                    handled = AsyncHttpServer.this.onRequest(this, res);
+                    if (handled)
+                        return;
+
+                    if (requestCallback == null) {
                         res.code(404);
                         res.end();
                         return;
                     }
 
-                    if (!getBody().readFullyOnRequest()) {
-                        onRequest(match, this, res);
-                    }
-                    else if (requestComplete) {
-                        onRequest(match, this, res);
-                    }
+                    if (!getBody().readFullyOnRequest() || requestComplete)
+                        onRequest();
                 }
 
                 @Override
                 public void onCompleted(Exception e) {
-                    // if the protocol was switched off http, ignore this request/response.
-                    if (res.code() == 101)
+                    if (isSwitchingProtocols(res))
                         return;
+
                     requestComplete = true;
                     super.onCompleted(e);
                     // no http pipelining, gc trashing if the socket dies
@@ -179,16 +206,27 @@ public class AsyncHttpServer {
                             mSocket.close();
                         }
                     });
+
+                    if (e != null) {
+                        mSocket.close();
+                        return;
+                    }
+
                     handleOnCompleted();
 
-                    if (getBody().readFullyOnRequest()) {
-                        onRequest(match, this, res);
+                    if (getBody().readFullyOnRequest() && !handled) {
+                        onRequest();
                     }
                 }
                 
                 private void handleOnCompleted() {
-                    if (requestComplete && responseComplete) {
-                        if (HttpUtil.isKeepAlive(Protocol.HTTP_1_1, getHeaders())) {
+                    // response may complete before request. the request may have a body, and
+                    // the response may be sent before it is fully sent.
+
+                    // if the protocol was switched off http, abandon the socket,
+                    // otherwise attempt to recycle it.
+                    if (requestComplete && responseComplete && !isSwitchingProtocols(res)) {
+                        if (isKeepAlive(self, res)) {
                             onAccepted(socket);
                         }
                         else {
@@ -208,6 +246,11 @@ public class AsyncHttpServer {
                     if (parts.length < 2)
                         return new Multimap();
                     return Multimap.parseQuery(parts[1]);
+                }
+
+                @Override
+                public String getUrl() {
+                    return fullPath;
                 }
             };
             req.setSocket(socket);
@@ -276,237 +319,6 @@ public class AsyncHttpServer {
     public CompletedCallback getErrorCallback() {
         return mCompletedCallback;
     }
-    
-    private static class Pair {
-        Pattern regex;
-        HttpServerRequestCallback callback;
-    }
-    
-    final Hashtable<String, ArrayList<Pair>> mActions = new Hashtable<String, ArrayList<Pair>>();
-
-    public void removeAction(String action, String regex) {
-        synchronized (mActions) {
-            ArrayList<Pair> pairs = mActions.get(action);
-            if (pairs == null)
-                return;
-            for (int i = 0; i < pairs.size(); i++) {
-                Pair p = pairs.get(i);
-                if (regex.equals(p.regex.toString())) {
-                    pairs.remove(i);
-                    return;
-                }
-            }
-        }
-    }
-    
-    public void addAction(String action, String regex, HttpServerRequestCallback callback) {
-        Pair p = new Pair();
-        p.regex = Pattern.compile("^" + regex);
-        p.callback = callback;
-        
-        synchronized (mActions) {
-            ArrayList<Pair> pairs = mActions.get(action);
-            if (pairs == null) {
-                pairs = new ArrayList<AsyncHttpServer.Pair>();
-                mActions.put(action, pairs);
-            }
-            pairs.add(p);
-        }
-    }
-
-    public static interface WebSocketRequestCallback {
-        public void onConnected(WebSocket webSocket, AsyncHttpServerRequest request);
-    }
-
-    public void websocket(String regex, final WebSocketRequestCallback callback) {
-        websocket(regex, null, callback);
-    }
-
-    public void websocket(String regex, final String protocol, final WebSocketRequestCallback callback) {
-        get(regex, new HttpServerRequestCallback() {
-            @Override
-            public void onRequest(final AsyncHttpServerRequest request, final AsyncHttpServerResponse response) {
-                boolean hasUpgrade = false;
-                String connection = request.getHeaders().get("Connection");
-                if (connection != null) {
-                    String[] connections = connection.split(",");
-                    for (String c: connections) {
-                        if ("Upgrade".equalsIgnoreCase(c.trim())) {
-                            hasUpgrade = true;
-                            break;
-                        }
-                    }
-                }
-                if (!"websocket".equalsIgnoreCase(request.getHeaders().get("Upgrade")) || !hasUpgrade) {
-                    response.code(404);
-                    response.end();
-                    return;
-                }
-                String peerProtocol = request.getHeaders().get("Sec-WebSocket-Protocol");
-                if (!TextUtils.equals(protocol, peerProtocol)) {
-                    response.code(404);
-                    response.end();
-                    return;
-                }
-                callback.onConnected(new WebSocketImpl(request, response), request);
-            }
-        });
-    }
-    
-    public void get(String regex, HttpServerRequestCallback callback) {
-        addAction(AsyncHttpGet.METHOD, regex, callback);
-    }
-    
-    public void post(String regex, HttpServerRequestCallback callback) {
-        addAction(AsyncHttpPost.METHOD, regex, callback);
-    }
-
-    public static android.util.Pair<Integer, InputStream> getAssetStream(final Context context, String asset) {
-        AssetManager am = context.getAssets();
-        try {
-            InputStream is = am.open(asset);
-            return new android.util.Pair<Integer, InputStream>(is.available(), is);
-        }
-        catch (IOException e) {
-            return null;
-        }
-    }
-
-    static Hashtable<String, String> mContentTypes = new Hashtable<String, String>();
-    {
-        mContentTypes.put("js", "application/javascript");
-        mContentTypes.put("json", "application/json");
-        mContentTypes.put("png", "image/png");
-        mContentTypes.put("jpg", "image/jpeg");
-        mContentTypes.put("html", "text/html");
-        mContentTypes.put("css", "text/css");
-        mContentTypes.put("mp4", "video/mp4");
-        mContentTypes.put("mov", "video/quicktime");
-        mContentTypes.put("wmv", "video/x-ms-wmv");
-    }
-
-    public static String getContentType(String path) {
-        String type = tryGetContentType(path);
-        if (type != null)
-            return type;
-        return "text/plain";
-    }
-
-    public static String tryGetContentType(String path) {
-        int index = path.lastIndexOf(".");
-        if (index != -1) {
-            String e = path.substring(index + 1);
-            String ct = mContentTypes.get(e);
-            if (ct != null)
-                return ct;
-        }
-        return null;
-    }
-
-    public void directory(Context context, String regex, final String assetPath) {
-        final Context _context = context.getApplicationContext();
-        addAction(AsyncHttpGet.METHOD, regex, new HttpServerRequestCallback() {
-            @Override
-            public void onRequest(AsyncHttpServerRequest request, final AsyncHttpServerResponse response) {
-                String path = request.getMatcher().replaceAll("");
-                android.util.Pair<Integer, InputStream> pair = getAssetStream(_context, assetPath + path);
-                if (pair == null || pair.second == null) {
-                    response.code(404);
-                    response.end();
-                    return;
-                }
-                final InputStream is = pair.second;
-                response.getHeaders().set("Content-Length", String.valueOf(pair.first));
-                response.code(200);
-                response.getHeaders().add("Content-Type", getContentType(assetPath + path));
-                Util.pump(is, response, new CompletedCallback() {
-                    @Override
-                    public void onCompleted(Exception ex) {
-                        response.end();
-                        StreamUtility.closeQuietly(is);
-                    }
-                });
-            }
-        });
-        addAction(AsyncHttpHead.METHOD, regex, new HttpServerRequestCallback() {
-            @Override
-            public void onRequest(AsyncHttpServerRequest request, final AsyncHttpServerResponse response) {
-                String path = request.getMatcher().replaceAll("");
-                android.util.Pair<Integer, InputStream> pair = getAssetStream(_context, assetPath + path);
-                if (pair == null || pair.second == null) {
-                    response.code(404);
-                    response.end();
-                    return;
-                }
-                final InputStream is = pair.second;
-                StreamUtility.closeQuietly(is);
-                response.getHeaders().set("Content-Length", String.valueOf(pair.first));
-                response.code(200);
-                response.getHeaders().add("Content-Type", getContentType(assetPath + path));
-                response.writeHead();
-                response.end();
-            }
-        });
-    }
-
-    public void directory(String regex, final File directory) {
-        directory(regex, directory, false);
-    }
-    
-    public void directory(String regex, final File directory, final boolean list) {
-        assert directory.isDirectory();
-        addAction("GET", regex, new HttpServerRequestCallback() {
-            @Override
-            public void onRequest(AsyncHttpServerRequest request, final AsyncHttpServerResponse response) {
-                String path = request.getMatcher().replaceAll("");
-                File file = new File(directory, path);
-                
-                if (file.isDirectory() && list) {
-                    ArrayList<File> dirs = new ArrayList<File>();
-                    ArrayList<File> files = new ArrayList<File>();
-                    for (File f: file.listFiles()) {
-                        if (f.isDirectory())
-                            dirs.add(f);
-                        else
-                            files.add(f);
-                    }
-                    
-                    Comparator<File> c = new Comparator<File>() {
-                        @Override
-                        public int compare(File lhs, File rhs) {
-                            return lhs.getName().compareTo(rhs.getName());
-                        }
-                    };
-                    
-                    Collections.sort(dirs, c);
-                    Collections.sort(files, c);
-                    
-                    files.addAll(0, dirs);
-                    
-                    return;
-                }
-                if (!file.isFile()) {
-                    response.code(404);
-                    response.end();
-                    return;
-                }
-                try {
-                    FileInputStream is = new FileInputStream(file);
-                    response.code(200);
-                    Util.pump(is, response, new CompletedCallback() {
-                        @Override
-                        public void onCompleted(Exception ex) {
-                            response.end();
-                        }
-                    });
-                }
-                catch (FileNotFoundException ex) {
-                    response.code(404);
-                    response.end();
-                }
-            }
-        });
-    }
 
     private static Hashtable<Integer, String> mCodes = new Hashtable<Integer, String>();
     static {
@@ -516,7 +328,11 @@ public class AsyncHttpServer {
         mCodes.put(101, "Switching Protocols");
         mCodes.put(301, "Moved Permanently");
         mCodes.put(302, "Found");
+        mCodes.put(304, "Not Modified");
+        mCodes.put(400, "Bad Request");
+        mCodes.put(401, "Unauthorized");
         mCodes.put(404, "Not Found");
+        mCodes.put(500, "Internal Server Error");
     }
     
     public static String getResponseCodeDescription(int code) {
@@ -525,4 +341,13 @@ public class AsyncHttpServer {
             return "Unknown";
         return d;
     }
+
+    public static void addResponseCodeDescription( int code, String description ) {
+        mCodes.put(code, description);
+    }
+
+    public static interface WebSocketRequestCallback {
+        void onConnected(WebSocket webSocket, AsyncHttpServerRequest request);
+    }
+
 }
