@@ -6,14 +6,20 @@ import com.koushikdutta.async.http.Headers;
 import com.koushikdutta.async.http.Multimap;
 import com.koushikdutta.async.http.body.AsyncHttpRequestBody;
 
+import java.util.Map;
 import java.util.regex.Matcher;
 
 public interface AsyncHttpServerRequest extends DataEmitter {
-    public Headers getHeaders();
-    public Matcher getMatcher();
-    public AsyncHttpRequestBody getBody();
-    public AsyncSocket getSocket();
-    public String getPath();
-    public Multimap getQuery();
-    public String getMethod();
+    Headers getHeaders();
+    Matcher getMatcher();
+    void setMatcher(Matcher matcher);
+    <T extends AsyncHttpRequestBody> T getBody();
+    AsyncSocket getSocket();
+    String getPath();
+    Multimap getQuery();
+    String getMethod();
+    String getUrl();
+
+    String get(String name);
+    Map<String, Object> getState();
 }
