@@ -9,16 +9,25 @@ import com.koushikdutta.async.callback.CompletedCallback;
 import com.koushikdutta.async.callback.DataCallback;
 import com.koushikdutta.async.http.Headers;
 import com.koushikdutta.async.http.HttpUtil;
+import com.koushikdutta.async.http.Multimap;
 import com.koushikdutta.async.http.Protocol;
 import com.koushikdutta.async.http.body.AsyncHttpRequestBody;
 
+import java.io.IOException;
+import java.util.HashMap;
+import java.util.Map;
 import java.util.regex.Matcher;
 
 public abstract class AsyncHttpServerRequestImpl extends FilteredDataEmitter implements AsyncHttpServerRequest, CompletedCallback {
     private String statusLine;
     private Headers mRawHeaders = new Headers();
     AsyncSocket mSocket;
-    Matcher mMatcher;
+    private HashMap<String, Object> state = new HashMap<>();
+
+    @Override
+    public HashMap<String, Object> getState() {
+        return state;
+    }
 
     public String getStatusLine() {
         return statusLine;
@@ -47,37 +56,41 @@ public abstract class AsyncHttpServerRequestImpl extends FilteredDataEmitter imp
     protected AsyncHttpRequestBody onUnknownBody(Headers headers) {
         return null;
     }
+    protected AsyncHttpRequestBody onBody(Headers headers) {
+        return null;
+    }
+
     
     StringCallback mHeaderCallback = new StringCallback() {
         @Override
         public void onStringAvailable(String s) {
-            try {
-                if (statusLine == null) {
-                    statusLine = s;
-                    if (!statusLine.contains("HTTP/")) {
-                        onNotHttp();
-                        mSocket.setDataCallback(null);
-                    }
+            if (statusLine == null) {
+                statusLine = s;
+                if (!statusLine.contains("HTTP/")) {
+                    onNotHttp();
+                    mSocket.setDataCallback(new NullDataCallback());
+                    report(new IOException("data/header received was not not http"));
                 }
-                else if (!"\r".equals(s)){
-                    mRawHeaders.addLine(s);
-                }
-                else {
-                    DataEmitter emitter = HttpUtil.getBodyDecoder(mSocket, Protocol.HTTP_1_1, mRawHeaders, true);
-//                    emitter.setEndCallback(mReporter);
-                    mBody = HttpUtil.getBody(emitter, mReporter, mRawHeaders);
-                    if (mBody == null) {
-                        mBody = onUnknownBody(mRawHeaders);
-                        if (mBody == null)
-                            mBody = new UnknownRequestBody(mRawHeaders.get("Content-Type"));
-                    }
-                    mBody.parse(emitter, mReporter);
-                    onHeadersReceived();
+
+                return;
+            }
+            if (!"\r".equals(s)){
+                mRawHeaders.addLine(s);
+                return;
+            }
+
+            DataEmitter emitter = HttpUtil.getBodyDecoder(mSocket, Protocol.HTTP_1_1, mRawHeaders, true);
+            mBody = onBody(mRawHeaders);
+            if (mBody == null) {
+                mBody = HttpUtil.getBody(emitter, mReporter, mRawHeaders);
+                if (mBody == null) {
+                    mBody = onUnknownBody(mRawHeaders);
+                    if (mBody == null)
+                        mBody = new UnknownRequestBody(mRawHeaders.get("Content-Type"));
                 }
             }
-            catch (Exception ex) {
-                onCompleted(ex);
-            }
+            mBody.parse(emitter, mReporter);
+            onHeadersReceived();
         }
     };
 
@@ -121,11 +134,6 @@ public abstract class AsyncHttpServerRequestImpl extends FilteredDataEmitter imp
         return mSocket.isChunked();
     }
 
-    @Override
-    public Matcher getMatcher() {
-        return mMatcher;
-    }
-
     AsyncHttpRequestBody mBody;
     @Override
     public AsyncHttpRequestBody getBody() {
@@ -152,5 +160,20 @@ public abstract class AsyncHttpServerRequestImpl extends FilteredDataEmitter imp
         if (mRawHeaders == null)
             return super.toString();
         return mRawHeaders.toPrefixString(statusLine);
+    }
+
+    @Override
+    public String get(String name) {
+        Multimap query = getQuery();
+        String ret = query.getString(name);
+        if (ret != null)
+            return ret;
+        AsyncHttpRequestBody body = getBody();
+        Object bodyObject = body.get();
+        if (bodyObject instanceof Multimap) {
+            Multimap map = (Multimap)bodyObject;
+            return map.getString(name);
+        }
+        return null;
     }
 }
