@@ -1,11 +1,9 @@
 package com.koushikdutta.async.parser;
 
-import com.koushikdutta.async.ByteBufferList;
 import com.koushikdutta.async.DataEmitter;
 import com.koushikdutta.async.DataSink;
 import com.koushikdutta.async.callback.CompletedCallback;
 import com.koushikdutta.async.future.Future;
-import com.koushikdutta.async.future.TransformFuture;
 import com.koushikdutta.async.http.body.DocumentBody;
 import com.koushikdutta.async.stream.ByteBufferListInputStream;
 
@@ -13,7 +11,6 @@ import org.w3c.dom.Document;
 
 import java.lang.reflect.Type;
 
-import javax.xml.parsers.DocumentBuilder;
 import javax.xml.parsers.DocumentBuilderFactory;
 
 /**
@@ -23,14 +20,7 @@ public class DocumentParser implements AsyncParser<Document> {
     @Override
     public Future<Document> parse(DataEmitter emitter) {
         return new ByteBufferListParser().parse(emitter)
-        .then(new TransformFuture<Document, ByteBufferList>() {
-            @Override
-            protected void transform(ByteBufferList result) throws Exception {
-                DocumentBuilderFactory dbf = DocumentBuilderFactory.newInstance();
-                DocumentBuilder db = dbf.newDocumentBuilder();
-                setComplete(db.parse(new ByteBufferListInputStream(result)));
-            }
-        });
+        .thenConvert(from -> DocumentBuilderFactory.newInstance().newDocumentBuilder().parse(new ByteBufferListInputStream(from)));
     }
 
     @Override
@@ -41,5 +31,10 @@ public class DocumentParser implements AsyncParser<Document> {
     @Override
     public Type getType() {
         return Document.class;
+    }
+
+    @Override
+    public String getMime() {
+        return "text/xml";
     }
 }
