@@ -4,7 +4,7 @@ import com.koushikdutta.async.DataEmitter;
 import com.koushikdutta.async.DataSink;
 import com.koushikdutta.async.callback.CompletedCallback;
 import com.koushikdutta.async.future.Future;
-import com.koushikdutta.async.future.TransformFuture;
+
 import org.json.JSONObject;
 
 import java.lang.reflect.Type;
@@ -15,13 +15,7 @@ import java.lang.reflect.Type;
 public class JSONObjectParser implements AsyncParser<JSONObject> {
     @Override
     public Future<JSONObject> parse(DataEmitter emitter) {
-        return new StringParser().parse(emitter)
-        .then(new TransformFuture<JSONObject, String>() {
-            @Override
-            protected void transform(String result) throws Exception {
-                setComplete(new JSONObject(result));
-            }
-        });
+        return new StringParser().parse(emitter).thenConvert(JSONObject::new);
     }
 
     @Override
@@ -32,5 +26,10 @@ public class JSONObjectParser implements AsyncParser<JSONObject> {
     @Override
     public Type getType() {
         return JSONObject.class;
+    }
+
+    @Override
+    public String getMime() {
+        return "application/json";
     }
 }
