@@ -1,11 +1,10 @@
 package com.koushikdutta.async.test;
 
 import android.net.Uri;
-import android.test.AndroidTestCase;
+import androidx.test.runner.AndroidJUnit4;
 import android.text.TextUtils;
 import android.util.Log;
 
-import com.google.android.gms.security.ProviderInstaller;
 import com.koushikdutta.async.AsyncServer;
 import com.koushikdutta.async.AsyncServerSocket;
 import com.koushikdutta.async.ByteBufferList;
@@ -28,9 +27,10 @@ import com.koushikdutta.async.http.server.AsyncHttpServerRequest;
 import com.koushikdutta.async.http.server.AsyncHttpServerResponse;
 import com.koushikdutta.async.http.server.AsyncProxyServer;
 
-import junit.framework.Assert;
-
 import org.json.JSONObject;
+import org.junit.Assert;
+import org.junit.Test;
+import org.junit.runner.RunWith;
 
 import java.io.File;
 import java.util.concurrent.CancellationException;
@@ -39,20 +39,22 @@ import java.util.concurrent.Semaphore;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.TimeoutException;
 
-import javax.net.ssl.SSLContext;
+import static androidx.test.InstrumentationRegistry.getContext;
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertFalse;
+import static org.junit.Assert.assertNotNull;
+import static org.junit.Assert.assertNull;
+import static org.junit.Assert.assertTrue;
+import static org.junit.Assert.fail;
 
-public class HttpClientTests extends AndroidTestCase {
-    AsyncHttpClient client;
+@RunWith(AndroidJUnit4.class)
+public class HttpClientTests {
     AsyncServer server = new AsyncServer();
-    
-    public HttpClientTests() {
-        super();
-        client = new AsyncHttpClient(server);
-    }
+    AsyncHttpClient client  = new AsyncHttpClient(server);
 
     @Override
-    protected void tearDown() throws Exception {
-        super.tearDown();
+    protected void finalize() throws Throwable {
+        super.finalize();
         client.getSSLSocketMiddleware().setConnectAllAddresses(false);
         client.getSocketMiddleware().setConnectAllAddresses(false);
         client.getSocketMiddleware().disableProxy();
@@ -104,11 +106,13 @@ public class HttpClientTests extends AndroidTestCase {
     */
 
     private static final long TIMEOUT = 10000L;
+    @Test
     public void testHomepage() throws Exception {
         Future<String> ret = client.executeString(new AsyncHttpGet("http://google.com"), null);
         assertNotNull(ret.get(TIMEOUT, TimeUnit.MILLISECONDS));
     }
 
+    @Test
     public void testClockworkMod() throws Exception {
         final Semaphore semaphore = new Semaphore(0);
         final Md5 md5 = Md5.createInstance();
@@ -268,8 +272,8 @@ public class HttpClientTests extends AndroidTestCase {
             public void onProgress(AsyncHttpResponse response, long downloaded, long total) {
                 semaphore.release();
             }
-        })
-        .setCallback(new FutureCallback<File>() {
+        });
+        fileFuture.setCallback(new FutureCallback<File>() {
             @Override
             public void onCompleted(Exception e, File result) {
                 assertTrue(e instanceof CancellationException);
