@@ -13,6 +13,8 @@ import com.koushikdutta.async.http.server.AsyncHttpServerRequest;
 
 import junit.framework.TestCase;
 
+import org.junit.Test;
+
 import java.util.concurrent.Semaphore;
 import java.util.concurrent.TimeUnit;
 
@@ -46,11 +48,12 @@ public class WebSocketTests extends TestCase {
         });
     }
     
-    private static final long TIMEOUT = 10000L; 
+    private static final long TIMEOUT = 60000L;
+    @Test
     public void testWebSocket() throws Exception {
         final Semaphore semaphore = new Semaphore(0);
 
-        AsyncHttpClient.getDefaultInstance().websocket("http://localhost:5000/ws", null, new WebSocketConnectCallback() {
+        AsyncHttpClient.getDefaultInstance().websocket("http://localhost:5000/ws", (String)null, new WebSocketConnectCallback() {
             @Override
             public void onCompleted(Exception ex, WebSocket webSocket) {
                 webSocket.send("hello");
